streamlit_app/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
openpyxl>=3.1.0
pillow>=10.0.0
scipy>=1.11.0
